langchain/requirements.txt
chainlit
langchain
chromadb
sentence-transformers
google-generativeai
python-dotenv
PyPDF2
pdfplumber
